fog: Remove GSI key includes
It's no longer required to install the compliance GSI
public keys into the ramdisk. Those public keys just need
to be included in the VTS test suite for GSI verification.

Bug: 149806769
Test: tree hugger
Those keys will be embedded into VtsSecurityAvb on host side
instead, to verify the GSI image used on the device.

Bug: 176869478
Test: TreeHugger
Change-Id: I3714de75b94be43a1364b89b92d893b336aa014b
Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: snnbyyds <snnbyyds@gmail.com>
Signed-off-by: Tanmsyffa <sltnmsyffaa@gmail.com>